requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cssselect==1.3.0
cssutils==2.11.1
distro==1.9.0
filelock==3.19.1
fsspec==2025.9.0
greenlet==3.2.4
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.11.0
keyring==25.6.0
llvmlite==0.44.0
lxml==6.0.1
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.108.0
openai-whisper==20250625
packaging==25.0
playsound3==3.2.6
playwright==1.55.0
pluggy==1.6.0
premailer==3.10.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pynput==1.8.1
pyobjc-core==11.1
pyobjc-framework-ApplicationServices==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreText==11.1
pyobjc-framework-Quartz==11.1
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-slugify==8.0.4
regex==2025.9.1
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
sympy==1.14.0
text-unidecode==1.3
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yagmail==0.15.293
